Integrate board into UI
